Move SearchResultPagination to containers
